Avoid nuint property for serializer compatibility

+++ src/Bonsai.ImGui/InputTextBuilder.cs
@@ -17,7 +17,7 @@ public class InputTextBuilder : TextControlBuilder<string>
     /// Gets or sets the maximum number of characters allowed in the text box.
     /// </summary>
     [Description("The maximum number of characters allowed in the text box.")]
-    public nuint Capacity { get; set; } = 1024;
+    public ulong Capacity { get; set; } = 1024;
 
     /// <inheritdoc/>
     protected override IObservable<string> Generate<TSource>(IObservable<TSource> source)
@@ -25,7 +25,7 @@ public class InputTextBuilder : TextControlBuilder<string>
         return Observable.Create<string>(observer =>
         {
             var buf = Text;
-            var bufSize = Capacity;
+            var bufSize = (nuint)Capacity;
             var label = $"##{Name ?? nameof(ImGui.InputText)}";
             var sourceObserver = Observer.Create<TSource>(
                 _ =>
